requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
fastapi>=0.110.0
uvicorn>=0.29.0
upstash-redis>=1.7.0
python-dotenv>=1.0.0
